Implemented algorithm to search checkmate in two

+++ src/Board.php
@@ -716,32 +716,34 @@ final class Board extends \SplObjectStorage
     public function undoMove(array $prevCastling): Board
     {
         $prev = end($this->history);
-        $piece = $this->getPieceByPosition($prev->move->position->next);
-        $this->detach($piece);
-        if ($prev->move->type === Move::PAWN_PROMOTES ||
-            $prev->move->type === Move::PAWN_CAPTURES_AND_PROMOTES) {
-            $pieceUndone = new Pawn($prev->move->color, $prev->position);
-        } else {
-            $pieceUndoneClass = new \ReflectionClass(get_class($piece));
-            $pieceUndone = $pieceUndoneClass->newInstanceArgs([
-                $prev->move->color,
-                $prev->position,
-                $piece->getIdentity() === Symbol::ROOK ? $piece->getType() : null, ]
-            );
+        if ($prev) {
+            $piece = $this->getPieceByPosition($prev->move->position->next);
+            $this->detach($piece);
+            if ($prev->move->type === Move::PAWN_PROMOTES ||
+                $prev->move->type === Move::PAWN_CAPTURES_AND_PROMOTES) {
+                $pieceUndone = new Pawn($prev->move->color, $prev->position);
+            } else {
+                $pieceUndoneClass = new \ReflectionClass(get_class($piece));
+                $pieceUndone = $pieceUndoneClass->newInstanceArgs([
+                    $prev->move->color,
+                    $prev->position,
+                    $piece->getIdentity() === Symbol::ROOK ? $piece->getType() : null, ]
+                );
+            }
+            $this->attach($pieceUndone);
+            if ($prev->move->isCapture && $capture = end($this->captures[$prev->move->color])) {
+                $capturedClass = new \ReflectionClass(Convert::toClassName($capture->captured->identity));
+                $this->attach($capturedClass->newInstanceArgs([
+                        $prev->move->color === Symbol::WHITE ? Symbol::BLACK : Symbol::WHITE,
+                        $capture->captured->position,
+                        $capture->captured->identity === Symbol::ROOK ? $capture->captured->type : null,
+                    ])
+                );
+                $this->popCapture($prev->move->color);
+            }
+            isset($prevCastling) ? $this->castling = $prevCastling : null;
+            $this->popHistory()->refresh();
         }
-        $this->attach($pieceUndone);
-        if ($prev->move->isCapture && $capture = end($this->captures[$prev->move->color])) {
-            $capturedClass = new \ReflectionClass(Convert::toClassName($capture->captured->identity));
-            $this->attach($capturedClass->newInstanceArgs([
-                    $prev->move->color === Symbol::WHITE ? Symbol::BLACK : Symbol::WHITE,
-                    $capture->captured->position,
-                    $capture->captured->identity === Symbol::ROOK ? $capture->captured->type : null,
-                ])
-            );
-            $this->popCapture($prev->move->color);
-        }
-        isset($prevCastling) ? $this->castling = $prevCastling : null;
-        $this->popHistory()->refresh();
 
         return $this;
     }
